Changed fore/background color merging in print_char
Changed the algorithm used for merging foreground and background colors
used in print_char.  The old algotihm used floats to allow for
partial transparency, but due to hard floats being disabled,
it was very slow.
The new algorithm instead uses a on/off transparency model,
such that any value above 127 means foreground, and any other means
background. Due to this not using floats it is much faster than
the old algorithm,