Reworked MesssageBox
MessageBox now handles it's own area, and does not use the entire area it's given.
MessageBox can also now have buttons, either none, Ok, or Confirm & Cancel.
The panic message box now has a red shade.

Also changed the interface for `display::add_widget` to be generic
and to not require it's argument to be boxed.

Co-authored-by: brynolv <k.jzfgxgh@gmail.com>
Co-authored-by: Vikdos <viktor.ofjall@gmail.com>